WIP - Adicionando serializadores
