+++ Oragon.RabbitMQ.sln
@@ -36,10 +36,16 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_", "_", "{7F63D8A9-AB50-4B
 		.editorconfig = .editorconfig
 		.gitattributes = .gitattributes
 		.gitignore = .gitignore
-		nuget.config = nuget.config
 		Global.Build.props = Global.Build.props
+		nuget.config = nuget.config
 		README.md = README.md
 	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Oragon.RabbitMQ.Serializer.NewtonsoftJson", "src\Oragon.RabbitMQ.Serializer.NewtonsoftJson\Oragon.RabbitMQ.Serializer.NewtonsoftJson.csproj", "{539B433C-3A59-4035-ACEC-81124E60FA69}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Oragon.RabbitMQ.Serializer.SystemTextJson", "src\Oragon.RabbitMQ.Serializer.SystemTextJson\Oragon.RabbitMQ.Serializer.SystemTextJson.csproj", "{BB40F5FE-54D7-4662-9DCE-FDC2EB7995BF}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Serializers", "Serializers", "{34C26882-DAE9-4D04-8EAD-506ED87BCFE7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -83,6 +89,14 @@ Global
 		{EE9412F2-E634-43B6-9DFB-4E6FFB4FDE2B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE9412F2-E634-43B6-9DFB-4E6FFB4FDE2B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EE9412F2-E634-43B6-9DFB-4E6FFB4FDE2B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{539B433C-3A59-4035-ACEC-81124E60FA69}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{539B433C-3A59-4035-ACEC-81124E60FA69}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{539B433C-3A59-4035-ACEC-81124E60FA69}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{539B433C-3A59-4035-ACEC-81124E60FA69}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BB40F5FE-54D7-4662-9DCE-FDC2EB7995BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BB40F5FE-54D7-4662-9DCE-FDC2EB7995BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BB40F5FE-54D7-4662-9DCE-FDC2EB7995BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BB40F5FE-54D7-4662-9DCE-FDC2EB7995BF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -98,6 +112,9 @@ Global
 		{1A369EC8-2C0C-42F7-9E37-2A3397F7D549} = {F36972F4-70F1-4315-8E0D-CA172765FF5B}
 		{CE5CB468-2D3F-4B8F-86E6-E72F6FDE5770} = {F36972F4-70F1-4315-8E0D-CA172765FF5B}
 		{EE9412F2-E634-43B6-9DFB-4E6FFB4FDE2B} = {F36972F4-70F1-4315-8E0D-CA172765FF5B}
+		{539B433C-3A59-4035-ACEC-81124E60FA69} = {34C26882-DAE9-4D04-8EAD-506ED87BCFE7}
+		{BB40F5FE-54D7-4662-9DCE-FDC2EB7995BF} = {34C26882-DAE9-4D04-8EAD-506ED87BCFE7}
+		{34C26882-DAE9-4D04-8EAD-506ED87BCFE7} = {1E36D008-C06F-4C9A-9712-6F2B3C34B2E6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {15BFB62B-2566-4748-95BA-300E2503F5B5}
