+++ src/Oragon.RabbitMQ.Abstractions/Serialization/IAMQPSerializer.cs
