+++ src/Oragon.RabbitMQ.Build.props
@@ -4,7 +4,7 @@
 
   <PropertyGroup>    
     <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
+    <Nullable>disable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <DebugType>embedded</DebugType>
     <DebugSymbols>true</DebugSymbols>
