--- src/Oragon.RabbitMQ.MinimalConsumer/Serialization/SystemTextJsonAMQPSerializer.cs
@@ -1,25 +1,0 @@
-﻿using RabbitMQ.Client;
-using System.Diagnostics;
-using System.Text;
-
-namespace Oragon.RabbitMQ.Serialization;
-
-
-public class SystemTextJsonAMQPSerializer : AMQPBaseSerializer
-{
-
-    public SystemTextJsonAMQPSerializer(ActivitySource activitySource) : base(activitySource, nameof(SystemTextJsonAMQPSerializer)) { }
-
-
-    protected override TResponse DeserializeInternal<TResponse>(IBasicProperties basicProperties, ReadOnlyMemory<byte> body)
-    {
-        var message = Encoding.UTF8.GetString(body.ToArray());
-        return System.Text.Json.JsonSerializer.Deserialize<TResponse>(message);
-    }
-
-    protected override byte[] SerializeInternal<T>(IBasicProperties basicProperties, T objectToSerialize)
-    {
-        return Encoding.UTF8.GetBytes(System.Text.Json.JsonSerializer.Serialize(objectToSerialize));
-    }
-}
-
