+++ src/Oragon.RabbitMQ.Serializer.SystemTextJson/Oragon.RabbitMQ.Serializer.SystemTextJson.csproj
@@ -1,0 +1,10 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="./../Oragon.RabbitMQ.Build.props" />
+
+	<ItemGroup>
+		<ProjectReference Include="..\Oragon.RabbitMQ.Abstractions\Oragon.RabbitMQ.Abstractions.csproj" />
+		<ProjectReference Include="..\Oragon.RabbitMQ\Oragon.RabbitMQ.csproj" />
+	</ItemGroup>
+
+</Project>
