--- src/Oragon.RabbitMQ/Class1.cs
@@ -1,12 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace Oragon.RabbitMQ;
-
-/// <summary>
-/// Teste
-/// </summary>
-public class Class1
-{
-
-}
